updated comments and admin routes
